input terminal

+++ Libraries/API-LIB/Header/ellips.h
@@ -10,6 +10,6 @@
 
 #include "includes.h"
 
-void draw_ellips(int x_center, int y_center, int x_rad, int y_rad, uint8_t color);
+void draw_ellips(int x_center, int y_center, int x_rad, int y_rad, int color);
 
 #endif /* API_LIB_HEADER_ELLIPS_H_ */
